Updated pylint to latest stable.
Since 1.8.0 pylint checks for functions/methods without an explicit
default return. Python always returns None for undefined conditionals
inside functions/methods, but since 1.8.0 pylint warns about doing so.

This "fixes" the warning, but not the source of this issue. We should
rework this a little bit later, so such thing are not needed.