Ask the OS for the correct local address
"Same network, fall back to default route" does not work if the
default route points to an external network but the gateway is within
another local subnet.

Thus, we simply fall back to asking the OS by simply connecting a socket
to the remote address and then reading the corresponding local address.

This removes our dependancy on the "netifaces" module.