Properly round KNX floats
KNX measurements frequently originate as decimals, i.e. a current of
0.22 A. We want to convert that back to the number which most accurately
represents that decimal, not the one that most accurately corresponds to
the math value of the float. Thus the incoming float gets rounded to 7
decimal digits. This does lose some precision, but KNX devices that are
more accurate than 6 digits don't exist.

For 2-byte floats, we round off at 5 digits to not lose any precision.