Fix exposure of date, time and datetime object when threaded=True (#925)
* Fix exposure of date, time and datetime object when threaded=True

I assume that the second event loop in the connection thread was executed before the
tasks in the task registry were registered. And probably for some reason the currently running event loop
was the one from the connection thread. Since the event loop had already been started with run_forever() no new
tasks would be scheduled.

* Add changelog