Split multicast socket for UDPTransport (#1678)
* Split multicast socket for UDPTransport

Create a separate socket when creating an UDPTransport with a remote
multicast address to receive and send messages.
This allows us to distinguish messages we sent ourselves.

Signed-off-by: Pieter De Gendt <pieter.degendt@basalte.be>

* Add integration test for router indications

Spawn multiple routers connections and verify we do not get our frames
echoed back.

Signed-off-by: Pieter De Gendt <pieter.degendt@gmail.com>

* Use the local address 127.0.0.1 to test routing indications

Try to keep the test bound to the loopback interface so no actual
frames are sent on the network when running pytest.

Signed-off-by: Pieter De Gendt <pieter.degendt@basalte.be>

* Use experimental multicast group during integration testing

Instead of binding to the loopback interface, use the experimental
multicast group address. See RFC 4727 for details.

Signed-off-by: Pieter De Gendt <pieter.degendt@basalte.be>

* add a comment for the mcast group

---------

Signed-off-by: Pieter De Gendt <pieter.degendt@basalte.be>
Signed-off-by: Pieter De Gendt <pieter.degendt@gmail.com>
Co-authored-by: farmio <farmio@alphart.net>